requirements.txt
easyocr==1.7.2
ollama==0.5.1
pdf2image==1.17.0
Pillow==11.3.0
reportlab==4.4.3
